requirements.txt
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
plotly==5.15.0
streamlit==1.24.1
chromadb==0.4.13
langchain==0.0.267
openai==0.27.8
python-dotenv==1.0.0
pydantic==1.10.12
streamlit-extras==0.3.0
streamlit-aggrid==0.3.4
streamlit-plotly-events==0.0.6
streamlit-elements==0.1.0
watchdog==3.0.0 